% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_od.R
\name{generate_fake_pc}
\alias{generate_fake_pc}
\title{generate_fake_pc - Generate Fake Passenger Counts}
\usage{
generate_fake_pc(N = 100, t_step = 300, S = 6, dest, boarders = 20)
}
\arguments{
\item{N}{the number of trips}

\item{S}{the number of stops}

\item{boarders}{The average number of boarder per stop.}
}
\value{
A matrix where each column is an OD vector
}
\description{
generate_fake_pc - Generate Fake Passenger Counts
}
