% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load}
\alias{load}
\title{load - No. passengers on the bus immediatly after each stops}
\usage{
load(x)
}
\arguments{
\item{x}{column vector of boardings and alightings}
}
\value{
vector of passengers loadings immediatly after each stops
}
\description{
load - No. passengers on the bus immediatly after each stops
}
