\name{odfrid-package}
\alias{odfrid-package}
\alias{odfrid}
\docType{package}
\title{
\packageTitle{odfrid}
}
\description{
\packageDescription{odfrid}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{odfrid}

Maintainer: \packageMaintainer{odfrid}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = odfrid)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{odfrid}}
%\section{Documentation Index}{\packageIndices{odfrid}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
