% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{routing_matrix}
\alias{routing_matrix}
\title{routing_matrix - construct routing matrix from no. stops (S)}
\usage{
routing_matrix(s)
}
\arguments{
\item{S}{length-one integer denoting the number of stops}
}
\value{
An (armadillo) integer matrix
}
\description{
routing_matrix - construct routing matrix from no. stops (S)
}
